Create WifiPassKey.py
This pull request introduces a new feature that allows users to view and retrieve the security key of a selected WiFi profile on Windows. The feature is implemented using Python's subprocess module and the netsh command-line tool.


- Added code to execute the netsh wlan show profiles command and extract a list of available WiFi profiles.
- Prompted the user to select a WiFi profile from the list.
- Executed the netsh wlan show profile "<selected_profile>" key=clear | findstr Key command to obtain the security key for the selected WiFi profile.
- Displayed the security key to the user.

Reason for the Pull Request:
This offers a convenient way for users to retrieve their WiFi network's security key, which can be useful in various scenarios, such as connecting new devices to the network. It simplifies the process by automating the command execution and presenting the user with a list of profiles to choose from.

How to Test

- Run the script.
- It will display a list of available WiFi profiles.
- Choose a WiFi profile by entering the corresponding number.
- The script will display the security key for the selected WiFi profile.